requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pypdf>=3.17.0
python-docx>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
pytest>=7.4.0
httpx>=0.25.0
